AIService/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# LangChain AI Agent
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
langgraph==0.0.20

# Vector Database
pgvector==0.2.4
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy>=1.24.3

# Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Caching
cachetools==5.3.2

# Scheduling for data sync
apscheduler==3.10.4

# Optional: For better text processing
# spacy==3.7.2
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz#egg=en_core_web_sm
